requirements.txt
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
coverage==5.1
crayons==0.3.1
cryptography==2.9.2
Django==3.0.6
django-annoying==0.10.6
django-cors-headers==3.3.0
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-stubs==1.2.0
freezegun==0.3.15
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
phonenumbers==8.12.6
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.1.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
webdriver-manager==3.1.0
wrapt==1.12.1
yapf==0.30.0
